+++ src/mono/wasm/debugger/tests/debugger-test-with-source-link/debugger-test-with-source-link.csproj
@@ -3,10 +3,9 @@
     <PackageId>LibraryTest</PackageId>
     <Version>1.0.0</Version>
     <PublishRepositoryUrl>true</PublishRepositoryUrl>
-    <RepositoryUrl>https://github.com/thaystg/runtime.git</RepositoryUrl>
+    <RepositoryUrl>https://github.com/dotnet/runtime.git</RepositoryUrl>
     <RepositoryType>git</RepositoryType>
     <EnableSourceLink>true</EnableSourceLink>
-    <RepositoryCommit>f53c3c8402b956126cccef81a3caa01815ef1161</RepositoryCommit>
     <IncludeSymbols>true</IncludeSymbols>
     <DebugType>embedded</DebugType>
     <IncludeSymbols>true</IncludeSymbols>
